Refactor simple_preemptor E2E test suite structure
